Give the errors a home that every module can import
The naming rules are about to leave plugin.py, and they raise
ValidationError. Were the exceptions to stay in plugin.py, that new module
would have to import the plugin, which imports it back. They move one floor
down, into the package itself, where btrfs.py already stands apart with its
own BtrfsError.

plugin.py imports them again, so the names it exposes today do not change.
